+++ onsiteprint-plugin/blocks/event/block-template-parts/participant-template.php
@@ -1,14 +1,14 @@
 <?php
 /* ------------------------------------------------------------------------
  *  Block Part Name: Participant Template
- ?  Updated: 2025-04-14 - 05:15 (Y:m:d - H:i)
+ ?  Updated: 2025-05-17 - 05:28 (Y:m:d - H:i)
 ---------------------------------------------------------------------------
  #  The Block Part Content
 --------------------------------------------------------------------------- */
 ?>
 
 <template id="<?= esc_attr($id) ?>-participant-template">
-    <article class="op-participant" data-validation="0" data-op-arrival="0" data-op-prints="0">
+    <article class="op-participant" data-validation="0" data-op-prints="0">
         <header>
             <p class="op-col-icon" data-icon="user">
                 <span class="op-icon" role="img" aria-label="User Icon"></span>
@@ -16,18 +16,26 @@
             <div class="op-col-lines">
                 <p class="op-col-line-1">
                     <span class="op-label">1</span>
-                    <span class="op-text">Thor</span>
+                    <span class="op-text"></span>
                 </p>
                 <p class="op-col-line-2">
                     <span class="op-label">2</span>
-                    <span class="op-text">Gud (Nordisk mytologi)</span>
+                    <span class="op-text"></span>
                 </p>
                 <p class="op-col-line-3">
                     <span class="op-label">3</span>
-                    <span class="op-text">Valhalla A/S</span>
+                    <span class="op-text"></span>
+                </p>
+                <p class="op-col-line-4">
+                    <span class="op-label">4</span>
+                    <span class="op-text"></span>
+                </p>
+                <p class="op-col-line-5">
+                    <span class="op-label">5</span>
+                    <span class="op-text"></span>
                 </p>
             </div>
-            <time class="op-col-arrival-time" datetime=""></time>
+            <time class="op-col-arrival-time" datetime=""><span class="op-text"></span></time>
             <div class="op-col-print-info">
                 <button class="op-participant-print op-button op-button-size-medium op-button-style-solid" data-color="primary-90" data-icon="print" data-icon-position="left">
                     <span class="op-icon" role="img" aria-label="Printer Icon"></span>